requirements.txt
# Зафиксированные версии для воспроизводимой установки (Python 3.10+).
# Установка: pip install -r bike_router/requirements.txt

numpy==2.3.5
geopandas==1.1.1
networkx==3.5
osmnx==2.0.7
# Нужен osmnx.distance.nearest_nodes для графа в WGS84 (иначе ImportError в Docker/проде)
scikit-learn==1.6.1
SRTM.py==0.3.7
shapely==2.1.2
requests==2.32.5
pillow==12.0.0
tqdm==4.67.1
python-dotenv==1.2.1
fastapi==0.124.4
uvicorn[standard]==0.38.0
pydantic==2.12.5
gunicorn==23.0.0
prometheus_client==0.21.1
